third_party/WebKit/LayoutTests/fast/events/mousemove-from-iframe-to-top-element-expected.txt

TOP

Tests that hover is lost from iframe when the mouse is moved from iframe to the element that stacked up on iframe

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS hoverBgColor is "rgb(255, 0, 0)"

PASS events[0] is "mouseenter"
PASS targetIds[0] is "target"

PASS events[1] is "mousemove"
PASS targetIds[1] is "target"

PASS bgColor is "rgb(255, 255, 0)"

PASS events[2] is "mouseleave"
PASS targetIds[2] is "target"

PASS events[3] is "mouseout"
PASS targetIds[3] is "target"
